pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "email-agent"
version = "0.1.0"
description = "AI-powered email assistant for executive email management"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    # Templates
    "jinja2>=3.1.0",
    # Auth & sessions
    "msal>=1.28.0",
    "cryptography>=42.0.0",
    "itsdangerous>=2.1.0",
    # HTTP client (replaces raw requests with async support)
    "httpx>=0.27.0",
    # LLM
    "anthropic>=0.40.0",
    # Config
    "pydantic-settings>=2.2.0",
    "pyyaml>=6.0",
    # Logging
    "python-json-logger>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-httpx>=0.30.0",
    "ruff>=0.4.0",
]

[tool.setuptools]
packages = ["app"]

[tool.setuptools.package-data]
app = ["templates/**/*", "static/**/*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"